src/harmonic_balancer.py
import numpy as np
import logging
from .circuit import QuantumResonanceCircuit
from .utils.utils import generate_harmony_vector
# from .utils.helpers import plot_convergence
from .system import System
from scipy.signal import find_peaks, butter, filtfilt, iirnotch
from scipy.optimize import minimize

class EnhancedHarmonicBalancer:
    def __init__(self, base_frequency: float, num_harmonics: int = 5, application: str = 'power'):
        self.base_frequency = base_frequency
        self.num_harmonics = num_harmonics
        self.application = application
        self.golden_ratio = (1 + np.sqrt(5)) / 2
        self.psi = np.random.uniform(0, 2*np.pi, num_harmonics)
        self.frequencies = np.array([base_frequency * (i + 1) for i in range(num_harmonics)])

        # Setup logging
        logging.basicConfig(level=logging.INFO)
        self.logger = logging.getLogger(__name__)

        # Initialize quantum circuit
        self.quantum_circuit = QuantumResonanceCircuit(num_harmonics)

        # Initialize history
        self.history = {'scores': [], 'states': []}

        # Set convergence threshold
        self.convergence_threshold = 1e-6

        # Initialize other attributes
        self.best_score = float('-inf')
        self.best_solution = None
        self.harmony_memory = []
        self.harmony_memory_size = 20
        self.max_iterations = 100
        self.num_qubits = num_harmonics  # Assuming num_qubits is equal to num_harmonics
    def check_convergence(self):
        # Check if the algorithm has converged
        if len(self.history['scores']) < 2:
            return False
        return abs(self.history['scores'][-1] - self.history['scores'][-2]) < self.convergence_threshold

    def generate_new_harmony(self, transition_constant):
       # Generate a new harmony vector based on the transition constant and quantum circuit
       base_harmony = generate_harmony_vector(self.num_qubits)
       quantum_state = self.quantum_circuit.initialize_state()
       evolved_state = self.quantum_circuit.evolve_state(quantum_state, transition_constant)
       quantum_influence = np.abs(evolved_state)**2
       return np.where(np.random.rand(self.num_qubits) < quantum_influence[:self.num_qubits], 1, base_harmony)

    def update_harmony_memory(self, new_vector, evolved_state, score):
        # Update the harmony memory with the new vector and score
        if score > self.best_score:
            self.best_score = score
            self.best_solution = new_vector
        self.harmony_memory.append(new_vector)
        if len(self.harmony_memory) > self.harmony_memory_size:
            self.harmony_memory.pop(0)

    def run_experiment(self):
        for iteration in range(self.max_iterations):
            new_harmony_vector = self.generate_new_harmony(transition_constant=0.1)
            evolved_state = self.quantum_circuit.evolve_state(new_harmony_vector, time=0.1)
            score = self.objective_function(evolved_state)
            self.update_harmony_memory(new_harmony_vector, evolved_state, score)
            self.history['scores'].append(score)
            self.history['states'].append(evolved_state)
            if self.check_convergence():
                break
        return self.best_solution, self.best_score

    def apply_quantum_resonance(self):
        total_possibilities = self.quantum_circuit.get_total_possibilities()
        return total_possibilities

    def resonance_condition(self, F0: float, k: float, m: float, omega: float, b: float) -> float:
        """Calculate amplitude at resonance condition."""
        return F0 / np.sqrt((k - m * omega**2)**2 + (b * omega)**2)

    def wave_interference(self, y1: np.ndarray, y2: np.ndarray) -> np.ndarray:
        """Calculate wave interference between two signals."""
        return y1 + y2

    def golden_harmony(self, R: float, F: float, E: float) -> float:
        """Calculate golden harmony metric."""
        return np.sqrt((R * F**2) + E**2)

    def detect_base_frequency(self, signal_data: np.ndarray, sample_rate: float) -> float:
        spectrum = np.abs(np.fft.fft(signal_data))
        freqs = np.fft.fftfreq(len(signal_data), 1/sample_rate)
        positive_spectrum = spectrum[:len(spectrum)//2]
        positive_freqs = freqs[:len(freqs)//2]

        peaks, _ = find_peaks(positive_spectrum, height=max(positive_spectrum)/10)

        if len(peaks) > 0:
            sorted_peaks = sorted(peaks, key=lambda x: positive_spectrum[x], reverse=True)
            for peak in sorted_peaks[:3]:
                detected_freq = positive_freqs[peak]
                if 0.8 * self.base_frequency <= detected_freq <= 1.2 * self.base_frequency:
                    if abs(detected_freq - self.base_frequency) > 0.1:
                        self.base_frequency = detected_freq
                        self.frequencies = np.array([self.base_frequency * (i + 1) for i in range(self.num_harmonics)])
                    return detected_freq

        return self.base_frequency

    def optimize_psi(self, signal_data: np.ndarray, sample_rate: float) -> float:
        def objective(psi):
            balanced = self.apply_psi(signal_data, psi, sample_rate)
            thd = self.calculate_thd(balanced, sample_rate)
            harmony = self.golden_harmony(thd, self.base_frequency, np.mean(np.abs(balanced)))
            return abs(harmony - self.golden_ratio)

        result = minimize(objective, self.psi, method='BFGS')
        self.psi = result.x
        return result.fun

    def apply_psi(self, signal_data: np.ndarray, psi: np.ndarray, sample_rate: float) -> np.ndarray:
        t = np.arange(len(signal_data)) / sample_rate
        correction = np.zeros_like(signal_data)
        for i, freq in enumerate(self.frequencies):
            correction += self.resonance_condition(1, 1, 1, 2*np.pi*freq, 0.1) * np.sin(2 * np.pi * freq * t + psi[i])
        return self.wave_interference(signal_data, -correction)

    def calculate_thd(self, signal_data: np.ndarray, sample_rate: float) -> float:
        spectrum = np.abs(np.fft.fft(signal_data))
        freqs = np.fft.fftfreq(len(signal_data), 1/sample_rate)
        fundamental_idx = np.argmax(spectrum[:len(spectrum)//2])
        harmonics = spectrum[fundamental_idx*2:fundamental_idx*(self.num_harmonics+1)]
        return np.sqrt(np.sum(harmonics**2)) / spectrum[fundamental_idx]

    def balance_signal(self, signal_data: np.ndarray, sample_rate: float) -> np.ndarray:
        detected_freq = self.detect_base_frequency(signal_data, sample_rate)

        # Update base_frequency if the detected frequency is significantly different
        if abs(detected_freq - self.base_frequency) > 0.1:  # You can adjust this threshold
            self.base_frequency = detected_freq
            self.frequencies = np.array([self.base_frequency * (i + 1) for i in range(self.num_harmonics)])
            self.logger.info(f"Base frequency updated to {self.base_frequency} Hz")

        self.optimize_psi(signal_data, sample_rate)
        balanced = self.apply_psi(signal_data, self.psi, sample_rate)

        # Apply application-specific processing
        if self.application == 'power':
            balanced = self.power_specific_processing(balanced, sample_rate)
        elif self.application == 'vibration':
            balanced = self.vibration_specific_processing(balanced, sample_rate)

        return balanced

    def power_specific_processing(self, signal_data: np.ndarray, sample_rate: float) -> np.ndarray:
        # Apply quantum entanglement simulation
        entanglement_effect = self.quantum_entanglement_simulation(self.num_harmonics)
        # Apply a series of notch filters to remove specific harmonics
        for harmonic in range(2, self.num_harmonics + 1):
            notch_freq = harmonic * self.base_frequency
            q = 30.0  # Quality factor
            w0 = notch_freq / (sample_rate / 2)
            b, a = iirnotch(w0, q)
            signal_data = filtfilt(b, a, signal_data)

        # Apply quantum influence
        quantum_influence = self.apply_quantum_resonance()
        signal_data *= (1 + 0.1 * quantum_influence)  # Adjust the scaling factor as needed

        return signal_data


    def vibration_specific_processing(self, signal_data: np.ndarray, sample_rate: float) -> np.ndarray:
        # Implement a simple low-pass filter to reduce high-frequency components
        cutoff_freq = 2 * self.base_frequency  # Adjust as needed
        nyquist = 0.5 * sample_rate
        normal_cutoff = cutoff_freq / nyquist
        b, a = butter(4, normal_cutoff, btype='low', analog=False)
        return filtfilt(b, a, signal_data)

    def quantum_entanglement_simulation(self, num_harmonics):
        # Implement a simple quantum entanglement simulation
        # This is a placeholder implementation and should be replaced with actual quantum simulation logic
        return np.random.rand(num_harmonics)

    def mri_harmonic_suppression(self, signal_data: np.ndarray, sample_rate: float) -> np.ndarray:
        """Suppress specific harmonics for MRI application."""
        for harmonic in [3, 5, 7]:  # Suppress 3rd, 5th, and 7th harmonics
            notch_freq = harmonic * self.base_frequency
            w0 = notch_freq / (sample_rate / 2)
            b, a = iirnotch(w0, 30)
            signal_data = filtfilt(b, a, signal_data)
        return signal_data
